fastapi-project/pyproject.toml
[tool.poetry]
name = "fastapi_project"
version = "0.1.0"
description = "A FastAPI project"
authors = ["Johnson <johnsonkummari@gmail.com>"]
license = "MIT"

packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.0"
email-validator = "^2.2.0"
motor = "^3.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
